Try setting up Zoom
One of the things that I found was that some very large diagrams (or
even some medium ones) would not not look great and you would want to
zoom in on a subsection of them.

This commit try setting up d3-zoom which provides a nice interface for
this behaviour. This unfortunately interferes with the click behaviour
of mermaid.

This commit moves the scripts into their own file.  Previously after the
introduction of a `css` file being passed in and the block becoming a
"format macro template" the `{` and `}` needed to be escaped.

This commit also moves some of the `use`s due `cargo fmt` being run
(ideally this should also be checked in CI).