Highlight the selected story.
There wasn't a good way to tell what story was selected. This made
clicking an element harder to use because it wasn't clear if it was
already selected.

This is highlighting a `rect` element. because of that we were limited
for which styles we could apply. Mermaid JS also uses `!important` which
made it too hard to set the `stroke`. Instead here we choose to use a
`filter` drop shadow because it works on svg nodes.

(the updates to `epic_info.rs` is related to running autoformatting)